lotsa docs and tests cleanup.
